+++ src/data/protocols/db/account/add-account-repository.ts
@@ -1,4 +1,4 @@
-import { AddAccountModel } from '@/domain/usecases/add-account'
+import { AddAccountModel } from '@/domain/usecases/account/add-account'
 import { AccountModel } from '@/domain/models/account'
 
 export interface AddAccountRepository {
