+++ src/data/usecases/account/authentication/db-authentication-protocols.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 export * from '@/domain/models/account'
-export * from '@/domain/usecases/authentication'
+export * from '@/domain/usecases/account/authentication'
 export * from '@/domain/models/authentication'
 export * from '@/data/protocols/db/account/load-account-by-email-repository'
 export * from '@/data/protocols/db/account/update-access-token-repository'
