+++ src/data/usecases/account/load-account-by-token/db-load-account-by-token-protocols.ts
@@ -2,4 +2,4 @@
 export * from '@/data/protocols/db/account/load-account-by-token-repository'
 export * from '@/domain/models/account'
 export * from '@/data/protocols/criptography/decrypter'
-export * from '@/domain/usecases/load-account-by-token'
+export * from '@/domain/usecases/account/load-account-by-token'
