+++ src/data/usecases/survey-result/save-survey-result/db-save-survey-result-protocols.ts
@@ -1,0 +1,4 @@
+/* istanbul ignore file */
+export * from '@/data/protocols/db/survey-result/save-survey-result-repository'
+export * from '@/domain/models/survey-result'
+export * from '@/domain/usecases/survey-result/save-survey-result'
