+++ src/data/usecases/survey/add-survey/db-add-survey-protocols.ts
@@ -1,3 +1,3 @@
 /* istanbul ignore file */
-export * from '@/domain/usecases/add-survey'
+export * from '@/domain/usecases/survey/add-survey'
 export * from '@/data/protocols/db/survey/add-survey-repository'
