+++ src/data/usecases/survey/load-survey-by-id/db-load-survey-by-id-protocols.ts
@@ -1,4 +1,4 @@
 /* istanbul ignore file */
 export * from '@/domain/models/survey'
 export * from '@/data/protocols/db/survey/load-survey-by-id-repository'
-export * from '@/domain/usecases/load-survey-by-id'
+export * from '@/domain/usecases/survey/load-survey-by-id'
