+++ src/domain/usecases/account/authentication.ts
@@ -1,4 +1,4 @@
-import { AuthenticationModel } from '../models/authentication'
+import { AuthenticationModel } from '../../models/authentication'
 
 export interface Authentication {
   auth: (authentication: AuthenticationModel) => Promise<string>
