+++ src/main/factories/usecases/account/authentication/db-authentication-factory.ts
@@ -2,8 +2,8 @@ import env from '../../../../config/env'
 import { BcryptAdapter } from '@/infra/criptography/bcrypt-adapter/bcrypt-adapter'
 import { JwtAdapter } from '@/infra/criptography/jwt-adapter/jwt-adapter'
 import { AccountMongoRepository } from '@/infra/db/mongodb/account/account-mongo-repository'
-import { DbAuthentication } from '@/data/usecases/authentication/db-authentication'
-import { Authentication } from '@/domain/usecases/authentication'
+import { DbAuthentication } from '@/data/usecases/account/authentication/db-authentication'
+import { Authentication } from '@/domain/usecases/account/authentication'
 
 export const makeDbAuthentication = (): Authentication => {
   const salt = 12
