+++ src/main/factories/usecases/account/load-account-by-token/db-load-account-by-token-factory.ts
@@ -1,6 +1,6 @@
 import { AccountMongoRepository } from './../../../../../infra/db/mongodb/account/account-mongo-repository'
-import { LoadAccountByToken } from './../../../../../domain/usecases/load-account-by-token'
-import { DbLoadAccountByToken } from '@/data/usecases/load-account-by-token/db-load-account-by-token'
+import { LoadAccountByToken } from '../../../../../domain/usecases/account/load-account-by-token'
+import { DbLoadAccountByToken } from '@/data/usecases/account/load-account-by-token/db-load-account-by-token'
 import { JwtAdapter } from '@/infra/criptography/jwt-adapter/jwt-adapter'
 import env from '../../../../config/env'
