+++ src/main/factories/usecases/survey/add-survey/db-add-survey-factory.ts
@@ -1,5 +1,5 @@
-import { AddSurvey } from '@/domain/usecases/add-survey'
-import { DbAddSurvey } from '@/data/usecases/add-survey/db-add-survey'
+import { AddSurvey } from '@/domain/usecases/survey/add-survey'
+import { DbAddSurvey } from '@/data/usecases/survey/add-survey/db-add-survey'
 import { SurveyMongoRepository } from '@/infra/db/mongodb/survey/survey-mongo-repository'
 
 export const makeDbAddsurvey = (): AddSurvey => {
