+++ src/presentation/controllers/login/login/login-controller-protocols.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 export * from '../../../protocols'
 export * from '@/validation/protocols/email-validator'
-export * from '@/domain/usecases/authentication'
+export * from '@/domain/usecases/account/authentication'
 export * from '@/domain/models/authentication'
 export * from '../../../protocols/validation'
