+++ src/presentation/controllers/login/signup/signup-controller-protocols.ts
@@ -1,8 +1,8 @@
 /* istanbul ignore file */
 export * from '../../../protocols'
 export * from '@/validation/protocols/email-validator'
-export * from '@/domain/usecases/add-account'
+export * from '@/domain/usecases/account/add-account'
 export * from '@/domain/models/account'
 export * from '@/domain/models/authentication'
 export * from '../../../protocols/validation'
-export * from '@/domain/usecases/authentication'
+export * from '@/domain/usecases/account/authentication'
