+++ src/presentation/controllers/survey-result/save-survey-result/save-survey-result-controller-protocols.ts
@@ -1,4 +1,4 @@
 /* istanbul ignore file */
 export * from '../../../protocols'
-export * from '@/domain/usecases/save-survey-result'
+export * from '@/domain/usecases/survey-result/save-survey-result'
 export * from '@/domain/models/survey-result'
