+++ src/presentation/controllers/survey/add-survey/add-survey-controller-protocols.ts
@@ -1,3 +1,3 @@
 /* istanbul ignore file */
 export * from '../../../protocols'
-export * from '@/domain/usecases/add-survey'
+export * from '@/domain/usecases/survey/add-survey'
