+++ src/presentation/controllers/survey/load-surveys/load-surveys-controller-protocols.ts
@@ -1,4 +1,4 @@
 /* istanbul ignore file */
 export * from '../../../protocols'
-export * from '@/domain/usecases/load-surveys'
+export * from '@/domain/usecases/survey/load-surveys'
 export * from '@/domain/models/survey'
