+++ src/presentation/middlewares/middleware-protocols.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
-export * from '@/domain/usecases/load-account-by-token'
+export * from '@/domain/usecases/account/load-account-by-token'
 export * from '../protocols/middleware'
 export * from '../protocols'
 export * from '@/domain/models/account'
